+++ src/datasources/results/types/QueryResults.types.ts
@@ -11,6 +11,24 @@ export interface QueryResults extends ResultsQuery {
   queryBy?: string;
 }
 
+export interface ResultsVariableQuery extends ResultsQuery {
+  properties?: string;
+  queryBy?: string;
+}
+
+export const ResultsVariableProperties = [
+  {
+    value: 'DATA_TABLE_IDS',
+    label: 'Data Table IDs',
+    description: 'Data Table IDs associated with the result',
+  },
+  {
+    value: 'PROGRAM_NAME',
+    label: 'Test Program Name',
+    description: 'Test Program Name of the result',
+  }
+]
+
 export const OrderBy = [
   {
     value: 'ID',